daemon: Extract machine store
We create the `Radicle.Daemon.MachineStore` module that provides a clear
interface for accessing machines stored in memory.

The module contains all code from `Radicle.Daemon.Common` which was the
definition of `Machine` and the `CachedMachines` hash map. The module
also contains the code from `Daemon` that modified the hash map of
machines.
